Project_Starter_Files-Building_a_Metrics_Dashboard/app/backend/requirements.txt
Flask==2.0.1
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-client
prometheus-flask-exporter
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation
opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift
jaeger-client==4.3.0
Flask-OpenTracing